add test for xlim behaviour for 1d plots
This was sparked by issue #149, which showed that 'kde' and 'fastkde'
were behaving slightly differently in applying xlim.

When the data spans the entire prior volume, the limits should be set to
the prior boundaries (e.g. uniform posterior to uniform prior). When the
data is contained in a much smaller volume than the prior volume (e.g.
Gaussian posterior to uniform prior), then the plotting limits should be
reduced accordingly.