bugfix for xlim for fastkde plots (#150)
* add test for xlim behaviour for 1d plots

This was sparked by issue #149, which showed that 'kde' and 'fastkde'
were behaving slightly differently in applying xlim.

When the data spans the entire prior volume, the limits should be set to
the prior boundaries (e.g. uniform posterior to uniform prior). When the
data is contained in a much smaller volume than the prior volume (e.g.
Gaussian posterior to uniform prior), then the plotting limits should be
reduced accordingly.

* change double call of check_bounds in fastkde_plot_1d to single call, record xmin and xmax the first time and pass through instead, otherwise there might be extra padding

* add test for limit for 2d plots analogous to 1d case

* change double call of check_bounds for fastkde_contour_plot_2d to single call analogous to 1d case

* version bump to 2.0.0-beta.8