add test for 1d KDE posterior comparing low to high `ncompress` value
The current version of anesthetic compresses the data before passing it
to `kde=gaussian_kde` to save time.  However, `gaussian_kde` is not
actually the time consuming bit. Only the later _call_ to `kde()` is
very time consuming. It is better to pass the full (uncompressed)
dataset to `kde=gaussian_kde` and then call `kde()` only on the
compressed data. This won't take much longer, but will render much more
precise posteriors.
This test will fail with the current, imprecise version, but should then
pass later with the improved version.