create new cobaya example data with visible burn-in
For the upcoming additions to the `MCMCSamples` class of a
`remove_burn_in` and a `Gelman_Rubin` method we need good example data
that actually visibly require the removal of a burn in. The new example
data was generated with an artificially low proposal scale that ensures
a slow and hence visible burn-in process.