reducing the `LinAlgError` test
The Gelman--Rubin statistic only works for linearly independent
parameters. It will raise a `LinAlgError` otherwise. To test for
failure, we copied a parameter. With numpy this did not lead to reliable
failures, which is why we initially repeated the parameter thrice. Now,
with the switch to scipy, a single copy should be enough.