Use the scipy version of `linalg.cholesky` which consistently raises `LinAngError` with linearly dependent columns (#287)
* use scipy cholesky, which consistently gives linalg error

* bump version number

* reducing the `LinAlgError` test

The Gelman--Rubin statistic only works for linearly independent
parameters. It will raise a `LinAlgError` otherwise. To test for
failure, we copied a parameter. With numpy this did not lead to reliable
failures, which is why we initially repeated the parameter thrice. Now,
with the switch to scipy, a single copy should be enough.

* Revert "reducing the `LinAlgError` test"

This reverts commit da96aea26c2c4a32a4423e131184467a289cd6bc.

---------

Co-authored-by: lukashergt <lthergt@phas.ubc.ca>