Tidy up (#294)
* Updating to tidy-up

* Updated

* Version bump

* Updated the readthedocs to install all rather than extras

* Removed extra print statement

* Added more authors to pyproject.toml

* reduce test data size in `mcmc_df` in `test_weighted_pandas`

This significantly reduces the duration of the longest tests (from 60s
down to 20s for the longest locally). This is particularly useful for
local parallel testing where the duration of the longest test
significantly affects the overall test runtime.

* introduce `df_small` fixture to speed up `test_weighted_pandas`

We are currently using dataframes with shape (3000, 4) for some tests
where numerical accuracy is not necessary (e.g. when checking only for
existance of correct labels in plots). Plotting these large dataframes
takes a lot of time, so switching to smaller dataframes significantly
reduces testing runtime.

* revert some earlier changes now that we have `df_small` which is better at speeding up tests

* add missing columns to `wdf_small`

* parametrise slowest test in `test_samples` to speed up parallel testing

* Alphabetical ordering

---------

Co-authored-by: lukashergt <lthergt@phas.ubc.ca>